0.6.6
Much of the caching system is done. It will also run, but only the base
config is shown and no buttons should be pushed or bad things...

- 'E.db.ElvUI_Animations' is now set up like a table rather than an
array. Values end in '_Tab' in order to differentiate a table from a
scalar value in the table.
- 'E.Options.args.ElvUI_Animations.args' is now set up by the same key
names as 'E.db.ElvUI_Animations'
- Removed use of 'V' for variables that should not carry over between
sessions, but left traces for reference. Use '_AddonTable' instead.