0.7.9
Quick Commit to save hours of work. Will also release a version since
the add-on does function enough for beta testing.

Added:

- Functionality and optimizations for pre-existing features. Much faster
and cleaner

- Sliders/options show the default option instead of the current when
'Use Defaults' is enabled! (This is because of the Caching system)

Known Issues:

- MouseOver alpha has been disabled until it is re-implemented with the
caching system

- Deleting a default tab is not saved after a /reload. No known fix. The
default profile reloads this data since it finds that they key is nil